+++ bin/nanocl/src/commands/state.rs
@@ -1,6 +1,7 @@
 use std::fs;
 use std::collections::HashMap;
 
+use nanocld_client::stubs::job::JobPartial;
 use ntex::rt;
 use futures::StreamExt;
 use clap::{Arg, Command, ArgAction};
@@ -12,7 +13,9 @@ use bollard_next::service::HostConfig;
 
 use nanocl_error::io::{IoError, FromIo, IoResult};
 use nanocld_client::NanocldClient;
-use nanocld_client::stubs::state::{StateMeta, StateApplyQuery};
+use nanocld_client::stubs::state::{
+  StateMeta, StateApplyQuery, StateJob, StateStreamStatus,
+};
 use nanocld_client::stubs::cargo::{OutputKind, CargoLogQuery};
 use nanocld_client::stubs::cargo_config::{
   CargoConfigPartial, Config as ContainerConfig,
@@ -292,6 +295,51 @@ pub async fn log_cargo(
     futures.push(fut);
   }
   Ok(futures)
+}
+
+pub async fn log_jobs(
+  client: &NanocldClient,
+  jobs: Vec<JobPartial>,
+) -> IoResult<()> {
+  let mut futures = Vec::new();
+  for job in jobs {
+    let client = client.clone();
+    let fut = async move {
+      match client.logs_job(&job.name).await {
+        Err(err) => {
+          eprintln!("Cannot attach to job {}: {err}", &job.name);
+        }
+        Ok(mut stream) => {
+          while let Some(output) = stream.next().await {
+            let output = match output {
+              Ok(output) => output,
+              Err(e) => {
+                eprintln!("Error: {e}");
+                break;
+              }
+            };
+            let name = output.container_name;
+            let data = output.log.data;
+            match output.log.kind {
+              OutputKind::StdOut => {
+                print!("[{name}]: {data}");
+              }
+              OutputKind::StdErr => {
+                eprint!("[{name}]: {data}");
+              }
+              OutputKind::Console => {
+                print!("[{name}]: {data}");
+              }
+              _ => {}
+            }
+          }
+        }
+      }
+    };
+    futures.push(fut);
+  }
+  futures::future::join_all(futures).await;
+  Ok(())
 }
 
 /// ## Attach to cargoes
@@ -647,6 +695,7 @@ async fn exec_state_apply(
   let args = parse_build_args(&state_ref.data, opts.args.clone())?;
   let mut namespace = String::from("global");
   let mut cargoes = Vec::new();
+  let mut jobs: Vec<JobPartial> = Vec::new();
   let data = match state_ref.meta.kind.as_str() {
     "Deployment" | "Cargo" => {
       namespace = match state_ref.data.get("Namespace") {
@@ -673,6 +722,22 @@ async fn exec_state_apply(
       cargoes = hooked_cargoes.clone();
       yaml["Cargoes"] = serde_yaml::to_value(&hooked_cargoes)
         .map_err(|err| err.map_err_context(|| "Unable to convert to yaml"))?;
+      yaml
+    }
+    "Job" => {
+      let yaml: serde_yaml::Value = inject_data(
+        &state_ref.format,
+        &state_ref.raw,
+        &args,
+        &cli_conf.context,
+        &client,
+      )
+      .await?;
+      jobs = match yaml.get("Jobs") {
+        Some(jobs) => serde_yaml::from_value(jobs.clone())
+          .map_err(|err| err.map_err_context(|| "Unable to convert to yaml"))?,
+        None => Vec::new(),
+      };
       yaml
     }
     _ => {
@@ -722,6 +787,9 @@ async fn exec_state_apply(
       }
     }
   }
+  // for job in jobs {
+
+  // }
   let data = serde_json::to_value(&data).map_err(|err| {
     err.map_err_context(|| "Unable to create json payload for the daemon")
   })?;
@@ -735,9 +803,13 @@ async fn exec_state_apply(
     .await?;
   let multiprogress = MultiProgress::new();
   multiprogress.set_move_cursor(false);
+  let mut has_error = false;
   let mut layers: HashMap<String, ProgressBar> = HashMap::new();
   while let Some(res) = stream.next().await {
     let res = res?;
+    if res.status == StateStreamStatus::Failed {
+      has_error = true;
+    }
     utils::state::update_progress(&multiprogress, &mut layers, &res.key, &res);
   }
   if opts.follow {
@@ -747,6 +819,13 @@ async fn exec_state_apply(
       ..Default::default()
     };
     log_cargoes(&client, cargoes, &query).await?;
+    log_jobs(&client, jobs).await?;
+  }
+  if has_error {
+    return Err(IoError::invalid_data(
+      "Statefile",
+      "couldn't apply correctly",
+    ));
   }
   Ok(())
 }