+++ crates/nanocl_stubs/src/job.rs
@@ -1,6 +1,12 @@
 use serde::{Serialize, Deserialize};
 use bollard_next::container::Config;
 
+use crate::{node::NodeContainerSummary, cargo::OutputLog};
+
+/// ## Job
+///
+/// A job is a collection of containers to run in sequence as a single unit to act like a command
+///
 #[derive(Debug, Default, Clone, PartialEq)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
@@ -28,6 +34,10 @@ pub struct Job {
   pub containers: Vec<Config>,
 }
 
+/// ## Job partial
+///
+/// Job partial is used to create a new job
+///
 #[derive(Debug, Default, Clone, PartialEq)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
@@ -53,4 +63,68 @@ pub struct JobPartial {
   pub metadata: Option<serde_json::Value>,
   /// List of container to run
   pub containers: Vec<Config>,
+}
+
+/// ## Job inspect
+/// Is a detailed view of a job
+/// It contains all the information about the job
+/// It also contains the list of containers
+///
+#[derive(Clone, Debug)]
+#[cfg_attr(feature = "test", derive(Default))]
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "PascalCase"))]
+pub struct JobInspect {
+  /// Name of the job
+  pub name: String,
+  /// When the job have been created
+  pub created_at: chrono::NaiveDateTime,
+  /// When the job have been updated
+  pub updated_at: chrono::NaiveDateTime,
+  /// Secrets to load as environment variables
+  #[cfg_attr(
+    feature = "serde",
+    serde(skip_serializing_if = "Option::is_none")
+  )]
+  pub secrets: Option<Vec<String>>,
+  #[cfg_attr(
+    feature = "serde",
+    serde(skip_serializing_if = "Option::is_none")
+  )]
+  /// Metadata (user defined)
+  #[cfg_attr(feature = "utoipa", schema(value_type = HashMap<String, Any>))]
+  pub metadata: Option<serde_json::Value>,
+  /// Containers to run
+  pub containers: Vec<Config>,
+  /// Number of instances
+  pub instance_total: usize,
+  /// Number of instance that succeeded
+  pub instance_success: usize,
+  /// List of containers
+  pub instances: Vec<NodeContainerSummary>,
+}
+
+impl From<JobInspect> for JobPartial {
+  fn from(job: JobInspect) -> Self {
+    Self {
+      name: job.name,
+      secrets: job.secrets,
+      metadata: job.metadata,
+      containers: job.containers,
+    }
+  }
+}
+
+/// ## Job log output
+///
+/// Output of a job log
+///
+#[derive(Debug)]
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "PascalCase"))]
+pub struct JobLogOutput {
+  pub container_name: String,
+  pub log: OutputLog,
 }