+++ bin/nanocl/changelog.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- nanocl state apply --reload to skip diff check.
+- State apply --reload to skip diff check.
+- Cargo image import with progress bar.
+
 
 ## [0.11.0] - 2023-11-06